+++ packages/riverpod/lib/src/listenable.dart
@@ -17,6 +17,7 @@ class _Listener<T> {
 /// A listenable object used by [ProviderElementBase] as a mean to subscribe
 /// to subsets of the state exposed by a provider.
 @internal
+@optionalTypeArgs
 class ProxyElementValueNotifier<T> extends _ValueListenable<T> {
   /// Directly obtain the value exposed, gratefully handling cases where
   /// [result] is null or in error state.