+++ packages/riverpod/lib/src/notifier/base.dart
@@ -205,9 +205,8 @@ class NotifierProviderElement<NotifierT extends NotifierBase<T>, T>
 
   @override
   void visitChildren({
-    required void Function(ProviderElementBase<Object?> element) elementVisitor,
-    required void Function(ProxyElementValueNotifier<Object?> element)
-        notifierVisitor,
+    required void Function(ProviderElementBase element) elementVisitor,
+    required void Function(ProxyElementValueNotifier element) notifierVisitor,
   }) {
     super.visitChildren(
       elementVisitor: elementVisitor,