+++ packages/riverpod/lib/src/state_provider/base.dart
@@ -161,9 +161,8 @@ class StateProviderElement<T> extends ProviderElementBase<T>
 
   @override
   void visitChildren({
-    required void Function(ProviderElementBase<Object?> element) elementVisitor,
-    required void Function(ProxyElementValueNotifier<Object?> element)
-        notifierVisitor,
+    required void Function(ProviderElementBase element) elementVisitor,
+    required void Function(ProxyElementValueNotifier element) notifierVisitor,
   }) {
     super.visitChildren(
       elementVisitor: elementVisitor,