+++ packages/riverpod/test/framework/listen_test.dart
@@ -1469,8 +1469,11 @@ void main() {
       final container = createContainer();
       final listener = Listener<int>();
 
-      final sub =
-          container.listen(provider, listener.call, fireImmediately: true);
+      final sub = container.listen(
+        provider,
+        listener.call,
+        fireImmediately: true,
+      );
 
       verify(listener(null, 0)).called(1);
       verifyNoMoreInteractions(listener);