+++ website/i18n/it/docusaurus-plugin-content-docs/current/essentials/auto_dispose/raw_auto_dispose.dart
@@ -1,7 +1,0 @@
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-/* SNIPPET START */
-// Possiamo specificare autoDispose per abilitare la distruzione dello stato automatica.
-final provider = Provider.autoDispose<int>((ref) {
-  return 0;
-});