+++ website/i18n/ko/docusaurus-plugin-content-docs/current/migration/from_change_notifier/declaration/index.tsx
@@ -1,9 +1,0 @@
-import raw from "!!raw-loader!./raw.dart";
-import codegen from "!!raw-loader!./declaration.dart";
-
-export default {
-  raw,
-  hooks: raw,
-  codegen,
-  hooksCodegen: codegen,
-};