+++ website/i18n/zh-Hans/docusaurus-plugin-content-docs/current/concepts/about_codegen/provider_type/non_code_gen/stream_provider.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-/* SNIPPET START */
-final exampleProvider =
-    StreamProvider.autoDispose<String>((ref) async* {
-  yield 'foo';
-});