+++ website/i18n/zh-Hans/docusaurus-plugin-content-docs/current/essentials/testing/provider_to_mock/codegen.dart
@@ -1,9 +1,0 @@
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'codegen.g.dart';
-
-/* SNIPPET START */
-// 一个急于初始化的提供者程序。
-@riverpod
-Future<String> example(ExampleRef ref) async => 'Hello world';
-/* SNIPPET END */